refactorizacion en los nombres de las clases e interfaces
