+++ src/main/java/com/reto/plazoleta/domain/api/ICustomerServicePort.java
@@ -2,7 +2,7 @@ package com.reto.plazoleta.domain.api;
 
 import com.reto.plazoleta.domain.model.OrderModel;
 
-public interface IOrderServicePort {
+public interface ICustomerServicePort {
 
     OrderModel saveOrder(OrderModel orderModelRequest, String tokenWithPrefixBearer);
 }
